# openapi.yaml
openapi: 3.0.0
info:
  title: 'Star Citizen API'
  contact:
    email: foxftw@star-citizen.wiki
  version: 3.0.0
servers:
  -
    url: 'https://api.star-citizen.wiki'
paths:
...
  /api/manufacturers:
    get:
      tags:
        - In-Game
        - Manufacturers
      summary: 'In-Game Manufacturers Overview'
      description: 'Returns paginated manufacturers grouped by name with optional pagination.'
      operationId: 1f9d905a5e89dba1162f35428b92847c
      parameters:
        -
          $ref: '#/components/parameters/page'
        -
          $ref: '#/components/parameters/page_number'
        -
          $ref: '#/components/parameters/page_size'
      responses:
        '200':
          description: 'List of Manufacturers'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/manufacturer_link'
  '/api/manufacturers/{manufacturer}':
    get:
      tags:
        - In-Game
        - Manufacturers
      summary: 'In-Game Manufacturer Detail'
      description: 'Retrieve a manufacturer by name, UUID, or code together with its products.'
      operationId: cba9b0f51268f6c1d4309aa573d37a9c
      parameters:
        -
          name: manufacturer
          in: path
          required: true
          schema:
            description: 'Manufacturer name, uuid, or code'
            type: string
      responses:
        '200':
          description: 'A Manufacturer and its products'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/manufacturer'
  /api/manufacturers/search:
    post:
      tags:
        - In-Game
        - Manufacturers
        - Search
      summary: 'In-Game Manufacturer Search'
      description: 'Search manufacturers by name, UUID, or code with optional pagination.'
      operationId: 26ff6a6dbd3785225cf297e83cc1b306
      parameters:
        -
          $ref: '#/components/parameters/page'
        -
          $ref: '#/components/parameters/page_number'
        -
          $ref: '#/components/parameters/page_size'
      requestBody:
        description: 'Manufacturer name, uuid, or code'
        required: true
        content:
          application/json:
            schema:
              type: object
            example: '{"query": "Anvil"}'
      responses:
        '200':
          description: 'A List of matching Manufacturers'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/manufacturer_link'
...
  /api/comm-links:
    get:
      tags:
        - Comm-Links
        - RSI-Website
      summary: 'Comm-Links Overview'
      description: 'Returns paginated comm-links with optional includes, categories, series, and channel filters.'
      operationId: 2204c384d22cb7168f6dd5830328c4dd
      parameters:
        -
          $ref: '#/components/parameters/page'
        -
          $ref: '#/components/parameters/page_number'
        -
          $ref: '#/components/parameters/page_size'
        -
          $ref: '#/components/parameters/comm_link_includes'
        -
          name: 'filter[channel]'
          in: query
          schema:
            type: string
        -
          name: 'filter[series]'
          in: query
          schema:
            type: string
        -
          name: 'filter[category]'
          in: query
          schema:
            type: string
        -
          name: sort
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 'List of Comm-Links'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/comm_link'
...
  '/api/comm-links/{id}':
    get:
      tags:
        - Comm-Links
        - RSI-Website
      summary: 'Comm-Link Detail'
      description: 'Retrieve a single comm-link by ID with the requested related resources.'
      operationId: 0f21f5e85cfa4d0fe0076f0d4a0d0ffd
      parameters:
        -
          $ref: '#/components/parameters/comm_link_includes'
        -
          name: id
          in: path
          required: true
          schema:
            description: 'Comm-Link ID, starting from 12663'
            type: integer
            format: int64
            minimum: 12663
      responses:
        '200':
          description: 'A singular Comm-Link'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/comm_link'
        '404':
          description: 'No Comm-Link with specified ID found.'
...
components:
  schemas:
...
    manufacturer_link:
      title: 'Link to the detail page of a manufacturer'
      properties:
        name:
          type: string
        code:
          type: string
        uuid:
          type: string
        link:
          type: string
      type: object
    manufacturer:
      title: 'In game manufacturer'
      properties:
        name:
          type: string
        code:
          type: string
        uuid:
          type: string
      type: object
...
    comm_link_link:
      title: 'Comm-Link Link'
      description: 'Resource link to a Comm-Link'
      properties:
        api_url:
          type: string
      type: object
    comm_link:
      title: Comm-Link
      description: 'A Comm-Link'
      properties:
        id:
          type: integer
        title:
          type: string
        rsi_url:
          description: 'Link to Comm-Link on RSI Website.'
          type: string
        api_url:
          description: 'API Link to this resource'
          type: string
        api_public_url:
          type: string
        channel:
          type: string
        category:
          type: string
        series:
          type: string
        images:
          type: array
          items:
            $ref: '#/components/schemas/comm_link_image'
        images_count:
          type: integer
        links:
          type: array
          items:
            $ref: '#/components/schemas/comm_link_link'
        links_count:
          type: integer
        comment_count:
          type: integer
        created_at:
          type: string
        translations:
          $ref: '#/components/schemas/translation'
      type: object
    comm_link_image:
      title: 'Comm-Link Image'
      description: 'Image used in a Comm-Link'
      properties:
        rsi_url:
          type: string
        api_url:
          type: string
          nullable: true
        alt:
          type: string
        size:
          type: integer
        mime_type:
          type: string
        last_modified:
          type: string
        tags:
          type: array
          items:
            type: string
          nullable: true
        similar_url:
          type: string
      type: object
...
    translation:
      title: 'Grouped Translations'
      description: 'Translations of an entity'
      properties:
        en:
          type: string
        de:
          type: string
        zh:
          type: string
      type: object
...
  parameters:
    comm_link_includes:
      name: include
      in: query
      explode: false
      allowReserved: true
      schema:
        description: 'Available Comm-Link includes'
        type: array
        items:
          type: string
          enum:
            - translations
            - images
            - links
    page:
      name: page
      in: query
      description: 'Page number for pagination (starts at 1).'
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    page_number:
      name: 'page[number]'
      in: query
      description: 'Page number for pagination (starts at 1).'
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    page_size:
      name: 'page[size]'
      in: query
      description: 'Number of results to return per page.'
      required: false
      schema:
        type: integer
        default: 30
        maximum: 200
        minimum: 1
...